requirements.txt
keras
tensorflow
h5py
numpy
flask
pillow
gunicorn
wheel
boto3
requests
python-dotenv
gpiozero
python-daemon